+++ Sources/CocoaLumberjack/Extensions/DDFileLogger+Buffering.m
@@ -110,23 +110,27 @@ static NSUInteger DDGetDefaultBufferSizeBytes() {
 #pragma mark - Logging
 
 - (void)logMessage:(DDLogMessage *)logMessage {
+    // Don't need to check for isOnInternalLoggerQueue, -lt_dataForMessage: will do it for us.
     NSData *data = [_fileLogger lt_dataForMessage:logMessage];
-    NSUInteger length = data.length;
-    if (length == 0) {
+
+    if (data.length == 0) {
         return;
     }
 
-#ifndef DEBUG
-    __unused
+    [data enumerateByteRangesUsingBlock:^(const void * __nonnull bytes, NSRange byteRange, BOOL * __nonnull __unused stop) {
+        NSUInteger bytesLength = byteRange.length;
+#ifdef NS_BLOCK_ASSERTIONS
+        __unused
 #endif
-    NSInteger written = [_buffer write:[data bytes] maxLength:length];
-    NSAssert(written == (NSInteger)length, @"Failed to write to memory buffer.");
+        NSInteger written = [_buffer write:bytes maxLength:bytesLength];
+        NSAssert(written > 0 && (NSUInteger)written == bytesLength, @"Failed to write to memory buffer.");
 
-    _currentBufferSizeBytes += length;
+        _currentBufferSizeBytes += bytesLength;
 
-    if (_currentBufferSizeBytes >= _maxBufferSizeBytes) {
-        [self lt_sendBufferedDataToFileLogger];
-    }
+        if (_currentBufferSizeBytes >= _maxBufferSizeBytes) {
+            [self lt_sendBufferedDataToFileLogger];
+        }
+    }];
 }
 
 - (void)flush {